Give flaky tests some more time to succeed (#13)
Some of the tests in observability_test.go fail randomly.
It appears to be timing instability. This increases the pauses
temporarily to see if tests succeed to confirm it is indeed
a timing issue.

We will need a proper fix for the tests by replacing pauses with
waits for the right condition.